pyproject.toml
[project]
name = "practice-sequential-data"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ipykernel>=6.30.1",
    "loguru>=0.7.3",
    "matplotlib>=3.10.5",
    "numba>=0.61.2",
    "numpy>=2.2.6",
    "pandas>=2.3.1",
    "scikit-learn>=1.7.1",
    "shap>=0.48.0",
    "torchmetrics>=1.8.0",
    "xgboost==2.1.3",
]
